patch-6

+++ src/main/java/main/config/BotStartConfig.java
@@ -316,15 +316,15 @@ public class BotStartConfig {
                         .stream()
                         .collect(Collectors.toMap(Participants::getUserIdAsString, Participants::getUserIdAsString));
 
-                GiveawayRegistry.getInstance().putGift(
-                        guild_long_id,
-                        new Gift(guild_long_id,
-                                channel_long_id,
-                                id_user_who_create_giveaway,
-                                //Добавляем пользователей в hashmap
-                                participantsList,
-                                activeGiveawayRepository,
-                                participantsRepository));
+                Gift gift = new Gift(guild_long_id,
+                        channel_long_id,
+                        id_user_who_create_giveaway,
+                        //Добавляем пользователей в hashmap
+                        participantsList,
+                        activeGiveawayRepository,
+                        participantsRepository);
+
+                GiveawayRegistry.getInstance().putGift(guild_long_id, gift);
 
                 //Устанавливаем счетчик на верное число
                 GiveawayRegistry.getInstance().getGift(guild_long_id).setCount(participantsList.size());
@@ -345,7 +345,7 @@ public class BotStartConfig {
                     Date date = new Date(date_end_giveaway.getTime());
                     timer.schedule(stopGiveawayByTimer, date);
 
-                    GiveawayRegistry.getInstance().putGiveawayTimer(guild_long_id, timer);
+                    GiveawayRegistry.getInstance().putGiveawayTimer(guild_long_id, stopGiveawayByTimer, timer);
                 }
             }
             rs.close();
@@ -392,47 +392,48 @@ public class BotStartConfig {
 
                         //-1 because one Bot
                         if (hasGift(guildIdLong) && reactions != null && reactions.get(0).getCount() - 1 != gift.getListUsersSize()) {
-
                             for (MessageReaction reaction : reactions) {
-                                List<User> userList;
+                                Map<String, User> userList = reaction
+                                        .retrieveUsers()
+                                        .complete()
+                                        .stream()
+                                        .filter(user -> !user.isBot())
+                                        .filter(user -> !gift.hasUserInList(user.getId()))
+                                        .collect(Collectors.toMap(User::getId, user -> user));
 
                                 if (isForSpecificRole) {
-                                    Role roleGiveaway = jda.getRoleById(giveawayData.getRoleId());
-                                    userList = reaction
-                                            .retrieveUsers()
-                                            .complete()
-                                            .stream()
-                                            .filter(user -> !user.isBot())
-                                            .filter(user -> !gift.hasUserInList(user.getId()))
-                                            .filter(user -> guildById
-                                                    //TODO: This block thread may be use *parallelStream()*
-                                                    //TODO: RateLimit может быть в будущем
-                                                    .retrieveMember(user).complete()
-                                                    .getRoles()
-                                                    .contains(roleGiveaway))
-                                            .collect(Collectors.toList());
-                                } else {
-                                    userList = reaction
-                                            .retrieveUsers()
-                                            .complete()
-                                            .stream()
-                                            .filter(user -> !user.isBot())
-                                            .filter(user -> !gift.hasUserInList(user.getId()))
-                                            .collect(Collectors.toList());
+                                    try {
+                                        Map<String, User> userMapTemp = new HashMap<>(userList); //bad practice but it`s work
+                                        Role roleGiveaway = jda.getRoleById(giveawayData.getRoleId());
+                                        for (Map.Entry<String, User> entry : userMapTemp.entrySet()) {
+                                            Guild guild = jda.getGuildById(guildIdLong);
+                                            if (guild != null) {
+                                                guild.retrieveMember(entry.getValue())
+                                                        .onSuccess(member -> {
+                                                            boolean contains = member.getRoles().contains(roleGiveaway);
+                                                            if (!contains) {
+                                                                userList.remove(entry.getKey());
+                                                            }
+                                                        }).complete();
+                                            }
+                                        }
+                                    } catch (Exception e) {
+                                        e.printStackTrace();
+                                    }
                                 }
 
                                 //System.out.println("UserList count: " + userList);
                                 //Перебираем Users в реакциях
-                                for (User user : userList) {
+                                for (Map.Entry<String, User> entry : userList.entrySet()) {
                                     if (!hasGift(guildIdLong)) return;
-                                    gift.addUserToPoll(user);
+                                    gift.addUserToPoll(entry.getValue());
                                     //System.out.println("User id: " + user.getIdLong());
                                 }
                             }
                         }
                     }
                 } catch (Exception e) {
-                    if (e.getMessage().contains("10008: Unknown Message") || e.getMessage().contains("Missing permission: VIEW_CHANNEL")) {
+                    if (e.getMessage() != null && e.getMessage().contains("10008: Unknown Message") || e.getMessage().contains("Missing permission: VIEW_CHANNEL")) {
                         System.out.println("updateUserList() " + e.getMessage() + " удаляем!");
                         activeGiveawayRepository.deleteActiveGiveaways(guildIdLong);
                         GiveawayRegistry.getInstance().removeGuildFromGiveaway(guildIdLong);
@@ -440,6 +441,15 @@ public class BotStartConfig {
                         e.printStackTrace();
                     }
                 }
+            }
+            try {
+                Gift.GiveawayTimerStorage giveawayTimer = GiveawayRegistry.getInstance().getGiveawayTimer(guildIdLong);
+                StopGiveawayByTimer stopGiveawayByTimer = giveawayTimer.getStopGiveawayByTimer();
+                if (stopGiveawayByTimer.getCountDown() == 1) {
+                    stopGiveawayByTimer.countDown();
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
             }
         }
         Thread.sleep(2000L);
